requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
bitarray==3.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
ckzg==2.1.0
clickhouse-connect==0.8.15
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
cytoolz==1.0.1
eth-account==0.13.6
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
fonttools==4.56.0
frozenlist==1.5.0
git-filter-repo==2.47.0
greenlet==3.1.1
hexbytes==1.3.0
idna==3.10
kiwisolver==1.4.8
lz4==4.4.3
markdown-it-py==3.0.0
matplotlib==3.10.1
mdurl==0.1.2
multidict==6.2.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
parsimonious==0.10.0
pillow==11.1.0
propcache==0.3.1
psycopg2-binary==2.9.10
pycryptodome==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyunormalize==16.0.0
pywin32==310
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rlp==4.1.0
six==1.17.0
SQLAlchemy==2.0.40
toolz==1.0.0
tqdm==4.67.1
types-requests==2.32.0.20250306
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
web3==7.9.0
websockets==13.1
yarl==1.18.3
zstandard==0.23.0
